+++ src/test/java/es/in2/issuer/backend/shared/application/workflow/impl/CredentialSignerWorkflowImplTest.java
@@ -792,6 +792,8 @@ class CredentialSignerWorkflowImplTest {
         when(credentialProcedureService.getCredentialId(updatedProcedure)).thenReturn(Mono.just("cred-id-123"));
         when(procedureRetryService.handleInitialAction(any(UUID.class), eq(ActionType.UPLOAD_LABEL_TO_RESPONSE_URI), any()))
                 .thenReturn(Mono.empty());
+        when(credentialProcedureService.getCredentialSubjectId(updatedProcedure))
+                .thenReturn(Mono.just("product-spec-123"));
 
         StepVerifier.create(credentialSignerWorkflow.retrySignUnsignedCredential(processId, authorizationHeader, procedureId))
                 .verifyComplete();
@@ -813,6 +815,8 @@ class CredentialSignerWorkflowImplTest {
         CredentialProcedure updatedProcedure = mock(CredentialProcedure.class);
         when(updatedProcedure.getCredentialType()).thenReturn(LABEL_CREDENTIAL_TYPE);
         when(updatedProcedure.getEmail()).thenReturn("company@example.com");
+        when(credentialProcedureService.getCredentialSubjectId(updatedProcedure))
+                .thenReturn(Mono.just("product-spec-123"));
 
         when(accessTokenService.getCleanBearerToken(authorizationHeader)).thenReturn(Mono.just(token));
         when(backofficePdpService.validateSignCredential(processId, token, procedureId)).thenReturn(Mono.empty());
@@ -895,7 +899,8 @@ class CredentialSignerWorkflowImplTest {
         when(credentialProcedureService.getCredentialId(updatedProcedure)).thenReturn(Mono.just("cred-id-123"));
         when(procedureRetryService.handleInitialAction(any(UUID.class), eq(ActionType.UPLOAD_LABEL_TO_RESPONSE_URI), any()))
                 .thenReturn(Mono.empty());
-
+        when(credentialProcedureService.getCredentialSubjectId(updatedProcedure))
+                .thenReturn(Mono.just("product-spec-123"));
         StepVerifier.create(credentialSignerWorkflow.retrySignUnsignedCredential(processId, authorizationHeader, procedureId))
                 .verifyComplete();