+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v2.2.22](https://github.com/in2workspace/in2-issuer-api/releases/tag/v2.2.22)
+### Changed
+- If the upload of a label credential fails or all retry attempts are exhausted, send an email to the configured certifier and marketplace email addresses instead of the credential owner. If the upload succeeds after a retry, notify the configured certifier and marketplace email addresses as well. If the upload succeeds on the first attempt, only the credential owner is notified.
+- Show the product specification ID in messages related to label credential uploads.
+
 ## [v2.2.21](https://github.com/in2workspace/in2-issuer-api/releases/tag/v2.2.21)
 ### Changed
 - Modify LEARCredentialMachine issuance policy validations to enforce that non-admins can only include `ProductOffering`.